// out/model/token.ts
/**
 * TransferTo Calculator API
 * TransferTo Calculator
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface Token {
    /**
     * The unique identifier of a user
     */
    id?: string;
    /**
     * Access-token
     */
    access_token?: string;
    /**
     * Number of failed attempts
     */
    failed_attempts?: number;
    /**
     * Number of seconds before a new login is accepted
     */
    time_remaining?: number;
}